requirements.txt
nicegui == 2.19.*; python_version >= "3.8"
PyGithub == 2.6.*; python_version >= "3.8"
PyYAML == 6.0.*; python_version >= "3.8"
requests == 2.32.*; python_version >= "3.8"
types-PyYAML
types-requests

# bandit[baseline]:  # can be used to scan for additional security issues in the code `python -m bandit src/ -r -c .\.bandit`
# pip-system-certs:  # for issues with SSL certificates
